Get one time sorting working (and fix an issue with the sorting code that resulted in it sorting wrong).
The sort operations here are not necessarily efficient, I'd need a
better algorithm to find the minimum moves (or do it from the bottom?),
but I think that would also be less useful in the continuous case. Since
that's the main usecase I care about, I'm going to leave the inefficient
moves for now.

There can also be issues here if other users are sorting the playlist
while this tool is sorting things. I'm not really sure what to do about
that yet, as we have to use the updated snapshots we're given for our
operations, but I don't want to have to re-retrieve and re-calculate the
playlist between every operation... We'll see I guess.